Removed validators
