+++ src/commands/chain/height.ts
@@ -33,17 +33,15 @@ export default class extends ProfileCommand {
     execute(options: ProfileOptions) {
         this.spinner.start();
         const profile = this.getProfile(options);
-        const chainHttp = new ChainHttp(profile.url);
 
+        const chainHttp = new ChainHttp(profile.url);
         chainHttp.getBlockchainHeight().subscribe((height) => {
             this.spinner.stop(true);
             console.log(height.toString());
         }, (err) => {
             this.spinner.stop(true);
-            let text = '';
-            text += chalk.red('Error');
             err = err.message ? JSON.parse(err.message) : err;
-            console.log(text, err.body && err.body.message ? err.body.message : err);
+            console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
         });
     }
 }
