+++ src/commands/restriction/account.ts
@@ -22,13 +22,11 @@ import {command, metadata, option} from 'clime';
 import {AccountRestriction, AccountRestrictionFlags, RestrictionAccountHttp} from 'nem2-sdk';
 import {ProfileCommand, ProfileOptions} from '../../profile.command';
 import {AddressResolver} from '../../resolvers/address.resolver';
-import {AddressValidator} from '../../validators/address.validator';
 
 export class CommandOptions extends ProfileOptions {
     @option({
         flag: 'a',
         description: 'Account address.',
-        validator: new AddressValidator(),
     })
     address: string;
 }
@@ -73,9 +71,9 @@ export default class extends ProfileCommand {
         this.spinner.start();
 
         const profile = this.getProfile(options);
-        const restrictionHttp = new RestrictionAccountHttp(profile.url);
         const address = new AddressResolver().resolve(options, profile);
 
+        const restrictionHttp = new RestrictionAccountHttp(profile.url);
         restrictionHttp.getAccountRestrictions(address)
             .subscribe((accountRestrictions: any) => {
                 this.spinner.stop(true);
@@ -86,10 +84,8 @@ export default class extends ProfileCommand {
                 }
             }, (err: any) => {
                 this.spinner.stop(true);
-                let text = '';
-                text += chalk.red('Error');
                 err = err.message ? JSON.parse(err.message) : err;
-                console.log(text, err.body && err.body.message ? err.body.message : err);
+                console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
             });
     }
 }
