+++ src/monitor.transaction.command.ts
@@ -16,10 +16,8 @@
  *
  */
 import {option} from 'clime';
-import {Address} from 'nem2-sdk';
 import {ProfileCommand, ProfileOptions} from './profile.command';
 import {TransactionService} from './service/transaction.service';
-import {AddressValidator} from './validators/address.validator';
 
 /**
  * Base command class to listen the blockchain.
@@ -43,7 +41,6 @@ export class MonitorAddressOptions extends ProfileOptions {
     @option({
         flag: 'a',
         description: 'Account address.',
-        validator: new AddressValidator(),
     })
     address: string;
 }
